Changed it into an MVC 3 Portable Area
