+++ MvcMembership.sln
@@ -12,7 +12,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvcMembership.Tests", "MvcM
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleWebsite", "SampleWebsite\SampleWebsite.csproj", "{80C05049-D202-49CD-B9EA-9210429FC8FA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleWebsite.Tests", "SampleWebsite.Tests\SampleWebsite.Tests.csproj", "{498E569C-DE6D-4F83-8660-6E58F818D803}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UserAdministrationArea", "UserAdministrationArea\UserAdministrationArea.csproj", "{F970C0FE-4E55-4B58-8BDD-4551F0195590}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UserAdministrationArea.Tests", "UserAdministrationArea.Tests\UserAdministrationArea.Tests.csproj", "{498E569C-DE6D-4F83-8660-6E58F818D803}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -32,6 +34,10 @@ Global
 		{80C05049-D202-49CD-B9EA-9210429FC8FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{80C05049-D202-49CD-B9EA-9210429FC8FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{80C05049-D202-49CD-B9EA-9210429FC8FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F970C0FE-4E55-4B58-8BDD-4551F0195590}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F970C0FE-4E55-4B58-8BDD-4551F0195590}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F970C0FE-4E55-4B58-8BDD-4551F0195590}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F970C0FE-4E55-4B58-8BDD-4551F0195590}.Release|Any CPU.Build.0 = Release|Any CPU
 		{498E569C-DE6D-4F83-8660-6E58F818D803}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{498E569C-DE6D-4F83-8660-6E58F818D803}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{498E569C-DE6D-4F83-8660-6E58F818D803}.Release|Any CPU.ActiveCfg = Release|Any CPU
