--- SampleWebsite/Areas/UserAdministration/Models/UserAdministration/IndexViewModel.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Web.Security;
-using PagedList;
-
-namespace SampleWebsite.Areas.UserAdministration.Models.UserAdministration
-{
-	public class IndexViewModel
-	{
-		public IPagedList<MembershipUser> Users { get; set; }
-		public IEnumerable<string> Roles { get; set; }
-	}
-}
