--- SampleWebsite/Areas/UserAdministration/Models/UserAdministration/RoleViewModel.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Web.Security;
-
-namespace SampleWebsite.Areas.UserAdministration.Models.UserAdministration
-{
-	public class RoleViewModel
-	{
-		public string Role { get; set; }
-		public IEnumerable<MembershipUser> Users { get; set; }
-	}
-}
