--- SampleWebsite/Areas/UserAdministration/Views/Web.config
@@ -1,35 +1,0 @@
-﻿<?xml version="1.0"?>
-
-<configuration>
-  <system.web>
-    <httpHandlers>
-      <add path="*" verb="*" type="System.Web.HttpNotFoundHandler"/>
-    </httpHandlers>
-
-    <!--
-        Enabling request validation in view pages would cause validation to occur
-        after the input has already been processed by the controller. By default
-        MVC performs request validation before a controller processes the input.
-        To change this behavior apply the ValidateInputAttribute to a
-        controller or action.
-    -->
-    <pages
-        validateRequest="false"
-        pageParserFilterType="System.Web.Mvc.ViewTypeParserFilter, System.Web.Mvc, Version=2.0.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35"
-        pageBaseType="System.Web.Mvc.ViewPage, System.Web.Mvc, Version=2.0.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35"
-        userControlBaseType="System.Web.Mvc.ViewUserControl, System.Web.Mvc, Version=2.0.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35">
-      <controls>
-        <add assembly="System.Web.Mvc, Version=2.0.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35" namespace="System.Web.Mvc" tagPrefix="mvc" />
-      </controls>
-    </pages>
-  </system.web>
-
-  <system.webServer>
-    <validation validateIntegratedModeConfiguration="false" />
-    
-    <handlers>
-      <remove name="BlockViewHandler"/>
-      <add name="BlockViewHandler" path="*" verb="*" preCondition="integratedMode" type="System.Web.HttpNotFoundHandler" />
-    </handlers>
-  </system.webServer>
-</configuration>
