+++ SampleWebsite/Content/Site.css
@@ -3,78 +3,72 @@ The base color for this template is #5c87b2. If you'd like
 to use a different color start by replacing all instances of
 #5c87b2 with your new color.
 ----------------------------------------------------------*/
-body
-{
+body {
     background-color: #5c87b2;
-    font-size: .75em;
-    font-family: Verdana, Helvetica, Sans-Serif;
+    font-size: .85em;
+    font-family: "Trebuchet MS", Verdana, Helvetica, Sans-Serif;
     margin: 0;
     padding: 0;
     color: #696969;
 }
 
-a:link
-{
+a:link {
     color: #034af3;
     text-decoration: underline;
 }
-a:visited
-{
+
+a:visited {
     color: #505abc;
 }
-a:hover
-{
+
+a:hover {
     color: #1d60ff;
     text-decoration: none;
 }
-a:active
-{
+
+a:active {
     color: #12eb87;
 }
 
-p, ul
-{
+p, ul {
     margin-bottom: 20px;
     line-height: 1.6em;
 }
 
-/* HEADINGS   
-----------------------------------------------------------*/
-h1, h2, h3, h4, h5, h6
-{
-    font-size: 1.5em;
-    color: #000;
-    font-family: Arial, Helvetica, sans-serif;
+header,
+footer,
+nav,
+section {
+    display: block;
 }
 
-h1
-{
+/* HEADINGS   
+----------------------------------------------------------*/
+h1, h2, h3, h4, h5, h6 {
+    font-size: 1.5em;
+    color: #000;
+}
+
+h1 {
     font-size: 2em;
     padding-bottom: 0;
     margin-bottom: 0;
 }
-h2
-{
+
+h2 {
     padding: 0 0 10px 0;
 }
-h3
-{
+
+h3 {
     font-size: 1.2em;
 }
-h4
-{
+
+h4 {
     font-size: 1.1em;
 }
-h5, h6
-{
-    font-size: 1em;
-}
 
-/* this rule styles <h2> tags that are the 
-first child of the left and right table columns */
-.rightColumn > h1, .rightColumn > h2, .leftColumn > h1, .leftColumn > h2
-{
-    margin-top: 0;
+h5, h6 {
+    font-size: 1em;
 }
 
 /* PRIMARY LAYOUT ELEMENTS   
@@ -82,55 +76,55 @@ first child of the left and right table columns */
 
 /* you can specify a greater or lesser percentage for the 
 page width. Or, you can specify an exact pixel width. */
-.page
-{
+.page {
     width: 90%;
     margin-left: auto;
     margin-right: auto;
 }
 
-#header
-{
+header, #header {
     position: relative;
     margin-bottom: 0px;
     color: #000;
     padding: 0;
 }
 
-#header h1
-{
+header h1, #header h1 {
     font-weight: bold;
     padding: 5px 0;
     margin: 0;
     color: #fff;
     border: none;
     line-height: 2em;
-    font-family: Arial, Helvetica, sans-serif;
     font-size: 32px !important;
+    text-shadow: 1px 1px 2px #111;
 }
 
-#main
-{
+#main {
     padding: 30px 30px 15px 30px;
     background-color: #fff;
-    margin-bottom: 30px;
-    _height: 1px; /* only IE6 applies CSS properties starting with an underscore */
+    border-radius: 4px 0 0 0;
+    -webkit-border-radius: 4px 0 0 0;
+    -moz-border-radius: 4px 0 0 0;
 }
 
-#footer
-{
+footer, 
+#footer {
+    background-color: #fff;
     color: #999;
     padding: 10px 0;
     text-align: center;
     line-height: normal;
-    margin: 0;
+    margin: 0 0 30px 0;
     font-size: .9em;
+    border-radius: 0 0 4px 4px;
+    -webkit-border-radius: 0 0 4px 4px;
+    -moz-border-radius: 0 0 4px 4px;
 }
 
 /* TAB MENU   
 ----------------------------------------------------------*/
-ul#menu
-{
+ul#menu {
     border-bottom: 1px #5C87B2 solid;
     padding: 0 0 2px;
     position: relative;
@@ -138,14 +132,12 @@ ul#menu
     text-align: right;
 }
 
-ul#menu li
-{
+ul#menu li {
     display: inline;
     list-style: none;
 }
 
-ul#menu li#greeting
-{
+ul#menu li#greeting {
     padding: 10px 20px;
     font-weight: bold;
     text-decoration: none;
@@ -153,30 +145,29 @@ ul#menu li#greeting
     color: #fff;
 }
 
-ul#menu li a
-{
+ul#menu li a {
     padding: 10px 20px;
     font-weight: bold;
     text-decoration: none;
     line-height: 2.8em;
     background-color: #e8eef4;
     color: #034af3;
+    border-radius: 4px 4px 0 0;
+    -webkit-border-radius: 4px 4px 0 0;
+    -moz-border-radius: 4px 4px 0 0;
 }
 
-ul#menu li a:hover
-{
+ul#menu li a:hover {
     background-color: #fff;
     text-decoration: none;
 }
 
-ul#menu li a:active
-{
+ul#menu li a:active {
     background-color: #a6e2a6;
     text-decoration: none;
 }
 
-ul#menu li.selected a
-{
+ul#menu li.selected a {
     background-color: #fff;
     color: #000;
 }
@@ -184,161 +175,158 @@ ul#menu li.selected a
 /* FORM LAYOUT ELEMENTS   
 ----------------------------------------------------------*/
 
-fieldset 
-{
-    margin: 1em 0;
-    padding: 1em;
-    border: 1px solid #CCC;
+fieldset {
+    border: 1px solid #ddd;
+    padding: 0 1.4em 1.4em 1.4em;
+    margin: 0 0 1.5em 0;
 }
 
-fieldset p 
-{
-    margin: 2px 12px 10px 10px;
+legend {
+    font-size: 1.2em;
+    font-weight: bold;
 }
 
-legend 
-{
-    font-size: 1.1em;
-    font-weight: 600;
-    padding: 2px 4px 8px 4px;
+textarea {
+    min-height: 75px;
 }
 
-input[type="text"] 
-{
+input[type="text"], 
+input[type="password"] {
+    border: 1px solid #ccc;
+    padding: 2px;
+    font-size: 1.2em;
+    color: #444;
     width: 200px;
-    border: 1px solid #CCC;
 }
 
-input[type="password"] 
-{
+select {
+    border: 1px solid #ccc;
+    padding: 2px;
+    font-size: 1.2em;
+    color: #444;
     width: 200px;
-    border: 1px solid #CCC;
+}
+
+input[type="submit"] {
+    font-size: 1.2em;
+    padding: 5px;
 }
 
 /* TABLE
 ----------------------------------------------------------*/
 
-table 
-{
-  border: solid 1px #e8eef4;
-  border-collapse: collapse;
+table {
+    border: solid 1px #e8eef4;
+    border-collapse: collapse;
 }
 
-table td 
-{
-  padding: 5px;   
-  border: solid 1px #e8eef4;
+table td {
+    padding: 5px;
+    border: solid 1px #e8eef4;
 }
 
-table th
-{
-  padding: 6px 5px;
-  text-align: left;
-  background-color: #e8eef4; 
-  border: solid 1px #e8eef4;   
+table th {
+    padding: 6px 5px;
+    text-align: left;
+    background-color: #e8eef4;
+    border: solid 1px #e8eef4;
 }
 
 /* MISC  
 ----------------------------------------------------------*/
-.clear
-{
+.clear {
     clear: both;
 }
 
-.error
-{
-    color:Red;
+.error {
+    color: Red;
 }
 
-#menucontainer
-{
-    margin-top:40px;
+nav, 
+#menucontainer {
+    margin-top: 40px;
 }
 
-div#title
-{
-    display:block;
-    float:left;
-    text-align:left;
+div#title {
+    display: block;
+    float: left;
+    text-align: left;
 }
 
-#logindisplay
-{
-    font-size:1.1em;
-    display:block;
-    text-align:right;
-    margin:10px;
-    color:White;
+#logindisplay {
+    font-size: 1.1em;
+    display: block;
+    text-align: right;
+    margin: 10px;
+    color: White;
 }
 
-#logindisplay a:link
-{
+#logindisplay a:link {
     color: white;
     text-decoration: underline;
 }
 
-#logindisplay a:visited
-{
+#logindisplay a:visited {
     color: white;
     text-decoration: underline;
 }
 
-#logindisplay a:hover
-{
+#logindisplay a:hover {
     color: white;
     text-decoration: none;
 }
 
 /* Styles for validation helpers
 -----------------------------------------------------------*/
-.field-validation-error
-{
+.field-validation-error {
     color: #ff0000;
 }
 
-.field-validation-valid
-{
+.field-validation-valid {
     display: none;
 }
 
-.input-validation-error
-{
+.input-validation-error {
     border: 1px solid #ff0000;
     background-color: #ffeeee;
 }
 
-.validation-summary-errors
-{
+.validation-summary-errors {
     font-weight: bold;
     color: #ff0000;
 }
 
-.validation-summary-valid
-{
+.validation-summary-valid {
     display: none;
 }
 
 /* Styles for editor and display helpers
 ----------------------------------------------------------*/
-.display-label,
-.editor-label,
-.display-field,
-.editor-field
-{
-    margin: 0.5em 0;
+.display-label, 
+.editor-label {
+    margin: 1em 0 0 0;
 }
 
-.text-box
-{
+.display-field, 
+.editor-field {
+    margin: 0.5em 0 0 0;
+}
+
+.text-box {
     width: 30em;
 }
 
-.text-box.multi-line
-{
+.text-box.multi-line {
     height: 6.5em;
 }
 
-.tri-state
-{
+.tri-state {
     width: 6em;
+}
+
+/* */
+ul.paging > li{
+	float: left;
+	list-style-type: none;
+	margin: 1em;
 }
