+++ SampleWebsite/SampleWebsite.csproj
@@ -18,6 +18,7 @@
     <OldToolsVersion>3.5</OldToolsVersion>
     <UpgradeBackupLocation />
     <TargetFrameworkProfile />
+    <UseIISExpress>false</UseIISExpress>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -39,6 +40,9 @@
     <CodeAnalysisRuleSet>AllRules.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="MvcContrib">
+      <HintPath>..\packages\MvcContrib.Mvc3-ci.3.0.68.0\lib\MvcContrib.dll</HintPath>
+    </Reference>
     <Reference Include="PagedList, Version=1.1.0.0, Culture=neutral, PublicKeyToken=abbb863e9397c5e1, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
       <HintPath>..\References\PagedList - 1.1\PagedList.dll</HintPath>
@@ -54,7 +58,8 @@
     <Reference Include="System.Web.DynamicData" />
     <Reference Include="System.Web.Entity" />
     <Reference Include="System.Web.Extensions" />
-    <Reference Include="System.Web.Mvc, Version=2.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL" />
+    <Reference Include="System.Web.Helpers, Version=1.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL" />
+    <Reference Include="System.Web.Mvc, Version=3.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL" />
     <Reference Include="System.Web.Routing">
       <RequiredTargetFramework>3.5</RequiredTargetFramework>
     </Reference>
@@ -68,11 +73,6 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Areas\UserAdministration\Controllers\UserAdministrationController.cs" />
-    <Compile Include="Areas\UserAdministration\Models\UserAdministration\DetailsViewModel.cs" />
-    <Compile Include="Areas\UserAdministration\Models\UserAdministration\IndexViewModel.cs" />
-    <Compile Include="Areas\UserAdministration\Models\UserAdministration\RoleViewModel.cs" />
-    <Compile Include="Areas\UserAdministration\UserAdministrationAreaRegistration.cs" />
     <Compile Include="Controllers\AccountController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Default.aspx.cs">
@@ -86,11 +86,6 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="App_Data\blank.txt" />
-    <Content Include="Areas\UserAdministration\Views\UserAdministration\UsersRoles.aspx" />
-    <Content Include="Areas\UserAdministration\Views\UserAdministration\Password.aspx" />
-    <Content Include="Areas\UserAdministration\Views\UserAdministration\Details.aspx" />
-    <Content Include="Areas\UserAdministration\Views\UserAdministration\Index.aspx" />
-    <Content Include="Areas\UserAdministration\Views\UserAdministration\Role.aspx" />
     <Content Include="Content\MvcMembership.css" />
     <Content Include="Default.aspx" />
     <Content Include="Global.asax" />
@@ -113,19 +108,35 @@
     <Content Include="Views\Shared\Error.aspx" />
     <Content Include="Views\Shared\LogOnUserControl.ascx" />
     <Content Include="Views\Shared\Site.Master" />
-    <Content Include="Views\Web.config" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MvcMembership\MvcMembership.csproj">
       <Project>{C9D7F653-E589-44CE-9D41-89821F4F9E4E}</Project>
       <Name>MvcMembership</Name>
     </ProjectReference>
+    <ProjectReference Include="..\UserAdministrationArea\UserAdministrationArea.csproj">
+      <Project>{F970C0FE-4E55-4B58-8BDD-4551F0195590}</Project>
+      <Name>UserAdministrationArea</Name>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Models\" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="Areas\UserAdministration\Views\Web.config" />
+    <Content Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Views\Web.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Areas\Web.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Views\Shared\_Layout.cshtml" />
+    <Content Include="Views\Shared\_LogOnPartial.cshtml" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Views\_ViewStart.cshtml" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v10.0\WebApplications\Microsoft.WebApplication.targets" />
