+++ SampleWebsite/Views/Shared/_Layout.cshtml
@@ -1,0 +1,37 @@
+﻿<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <title>@ViewBag.Title</title>
+    <link href="@Url.Content("~/Content/Site.css")" rel="stylesheet" type="text/css" />
+    <script src="@Url.Content("~/Scripts/jquery-1.5.2.min.js")" type="text/javascript"></script>
+    <script src="@Url.Content("~/Scripts/modernizr-1.7.min.js")" type="text/javascript"></script>
+    @RenderSection("HeadSection", false)
+</head>
+<body>
+    <div class="page">
+        <header>
+            <div id="title">
+                <h1>My MVC Application</h1>
+            </div>
+            <div id="logindisplay">
+                @Html.Partial("_LogOnPartial")
+            </div>
+            <nav>
+                <ul id="menu">
+                    <li>@Html.ActionLink("Home", "Index", "Home", new { Area = "" }, new { })</li>
+                    <li>@Html.ActionLink("About", "About", "Home", new { Area = "" }, new { })</li>
+                    @if (Roles.IsUserInRole("Administrator")) { 
+                       <li>@Html.ActionLink("User Administration", "Index", "UserAdministration", new { Area = "UserAdministration" }, new { })</li>
+                    }
+                </ul>
+            </nav>
+        </header>
+        <section id="main">
+            @RenderBody()
+        </section>
+        <footer>
+        </footer>
+    </div>
+</body>
+</html>
