+++ SampleWebsite/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Mvc3Futures" version="3.0.20105.0" />
+  <package id="MvcContrib.Mvc3-ci" version="3.0.68.0" />
+</packages>
