+++ UserAdministrationArea/Models/RoleViewModel.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Web.Security;
+
+namespace UserAdministration.Models
+{
+	public class RoleViewModel
+	{
+		public string Role { get; set; }
+		public IEnumerable<MembershipUser> Users { get; set; }
+	}
+}
