+++ UserAdministrationArea/Views/UserAdministration/Role.cshtml
@@ -1,0 +1,29 @@
+﻿@using MvcContrib;
+@model UserAdministration.Models.RoleViewModel
+
+@{
+    ViewBag.Title = "Role: " + @Model.Role;
+}
+
+@section HeadSection {
+	<link href='@Url.Resource("Content.MvcMembership.css")' rel="stylesheet" type="text/css" />
+}
+
+<h2 class="mvcMembership">Role: @Model.Role</h2>
+
+<div class="mvcMembership-roleUsers">
+	@if(Model.Users.Count() > 0) { 
+		<ul class="mvcMembership">
+			@foreach(var user in Model.Users)  { 
+			    <li>
+				    @Html.ActionLink(user.UserName, "Details", new { id = user.ProviderUserKey })
+				    @using(Html.BeginForm("RemoveFromRole", "UserAdministration", new { id = user.ProviderUserKey, role = Model.Role })){ 
+					    <input type="submit" value="Remove From" />
+				    } 
+			    </li>
+			} 
+		</ul>
+	} else { 
+	    <p>No users are in this role.</p>
+	} 
+</div>
